RealEstatePricePrediction/modular_code/requirements.txt
fastapi
joblib
nltk
numpy
pandas
pydantic
scikit_learn
scipy
uvicorn
openpyxl
streamlit
streamlit-authenticator
mysql-connector-python
pandas
sqlalchemy
xlrd
yagmail
PyYAML
streamlit-aggrid


# fastapi==0.79.0
# joblib==1.1.0
# nltk==3.7
# numpy==1.23.1
# pandas==1.4.3
# pydantic==1.9.1
# scikit_learn==1.0.2
# scipy==1.9.0
# uvicorn==0.18.2
# openpyxl==3.0.10
# streamlit==1.29.0
# streamlit-authenticator==0.2.3
# mysql-connector-python==8.0.23
# pandas>=0.19
# sqlalchemy==1.3.19
# xlrd==1.2.0
# yagmail==0.14.256
# PyYAML==6.0
# streamlit-aggrid==0.3.4

